ml-model/templates/pytorch/{{cookiecutter.project_slug}}/requirements-inf.txt
mlflow
onnxruntime
